requirements.txt
Django==2.1.1
django-cors-headers==2.4.0
django-webpack-loader==0.6.0
djangorestframework==3.8.2
pytz==2018.5
